globalize hotels state using context
