+++ src/components/HotelsList/HotelsList.jsx
@@ -1,24 +1,17 @@
 import React from "react";
-import { Link, useSearchParams } from "react-router-dom";
-import useFetch from "../../hooks/useFetch";
+import { Link } from "react-router-dom";
+import { useHotels } from "../../contexts/HotelsProvider";
 import Loader from "../Loader/Loader";
 
 const HotelsList = () => {
-   const [searchParams, setSearchParams] = useSearchParams();
-   const destination = searchParams.get("destination");
-   const room = JSON.parse(searchParams.get("options"))?.room;
-
-   const { data, isLoading } = useFetch(
-      "http://localhost:5000/hotels",
-      `q=${destination || ""}&accommodates_gte=${room || 1}`,
-   );
+   const { hotels, isLoading } = useHotels();
 
    if (isLoading) return <Loader />;
 
    return (
       <div className="searchList">
-         <h2>Search Results ({data.length})</h2>
-         {data.map((item) => {
+         <h2>Search Results ({hotels.length})</h2>
+         {hotels.map((item) => {
             return (
                <Link
                   key={item.id}
